add TestNullReporter
