--- v1/tv/internal/traceview/traceview.go
@@ -1,9 +1,0 @@
-// Copyright (C) 2016 AppNeta, Inc. All rights reserved.
-
-package traceview
-
-// Determines if request should be traced, based on sample rate settings:
-// This is our only dependency on the liboboe C library.
-func shouldTraceRequest(layer, xtraceHeader string) (sampled bool, sampleRate, sampleSource int) {
-	return oboeSampleRequest(layer, xtraceHeader)
-}
